v1.3.1.2 -------- - Corrected credits type in sql file (26/03/2018) - Corrected french translation (26/03/2018)

+++ Resources/sql/purchase_credits.sql
@@ -16,7 +16,7 @@ SET FOREIGN_KEY_CHECKS=0;
 CREATE TABLE `user_transactions` (
   `id` bigint(20) unsigned NOT NULL AUTO_INCREMENT,
   `order_id` varchar(48) DEFAULT NULL,
-  `credits` mediumint(8) DEFAULT NULL,
+  `credits` smallint(5) DEFAULT NULL,
   `description` varchar(512) DEFAULT NULL,
   `user_id` bigint(20) unsigned DEFAULT NULL,
   `user_ip` varchar(48) DEFAULT NULL,
